Cgroup Resource sudo integration test
