+++ agent/engine/docker_task_engine_linux_test.go
@@ -1,4 +1,4 @@
-// +build linux,!integration
+// +build linux,!integration,!sudo
 
 // Copyright 2014-2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 //
