+++ agent/engine/docker_task_engine_unix_test.go
@@ -1,4 +1,4 @@
-// +build !windows,!integration
+// +build !windows,!integration,!sudo
 
 // Copyright 2014-2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 //
