+++ agent/engine/dockerstate/testutils/json_test.go
@@ -1,4 +1,4 @@
-// +build !integration
+// +build !integration,!sudo
 // Copyright 2014-2017 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License"). You may
