+++ agent/engine/task_manager_unix_test.go
@@ -1,4 +1,4 @@
-// +build linux,!integration
+// +build linux,!integration,!sudo
 // Copyright 2014-2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License"). You may
