+++ misc/taskmetadata-validator/build-in-docker
@@ -21,4 +21,4 @@ docker run \
 	-v ${AGENT_VENDOR_DIR}:/gopath/src \
 	-e CGO_ENABLED=0 \
 	-e GOPATH=/go:/gopath \
-	golang:1.7 go build -a -x -ldflags '-s' -o /out/taskmetadata-validator /in/taskmetadata-validator.go
+	golang:1.9 go build -a -x -ldflags '-s' -o /out/taskmetadata-validator /in/taskmetadata-validator.go
